LayoutTests/fast/dom/DocumentType/isEqualNode-expected.txt

Tests that isEqualNode() works as expected for DocumentType nodes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS htmlDocument.doctype.isEqualNode(htmlDoctype) is true
PASS successfullyParsed is true

TEST COMPLETE
